requirements.txt
flask==3.0.3
aiogram==3.6.0
aiosqlite==0.20.0
aiohttp==3.9.5
yookassa==3.6.0
requests==2.32.3

python-dotenv==1.0.1
